Formatting and code refactoring
