+++ prokress/src/config.ts
@@ -1,0 +1,5 @@
+// Development
+//export const API_URL = "http://localhost:8080";
+
+// Production
+export const API_URL = "https://project-management-app-prokress-backend.2.rahtiapp.fi";
