+++ prokress/src/pages/ProjectsPage.tsx
@@ -4,7 +4,7 @@ import { useState, useEffect } from 'react'
 import { getProjectsForUser } from '../helper/handler.tsx'
 import type { projectType } from '../helper/types.ts'
 import ContextMenu from '../Components/ContextMenu.tsx'
-
+import '../App.css'
 function ProjectsPage() {
     // contextMenu
     const [contextMenuMode, setContextMenuMode] = useState<"project" | "task" | null>(null);
@@ -16,7 +16,6 @@ function ProjectsPage() {
     const [sharedProjects, setSharedProjects] = useState<projectType[]>([]);
     const [isProjectDialogOpen, setProjectDialogOpen] = useState(false)
     const [creatingShared, setCreatingShared] = useState(false)
-    //const curDate = new Intl.DateTimeFormat('en-GB').format(new Date());
     const appUserId = localStorage.getItem('token')
 
     useEffect(() => {
@@ -43,13 +42,13 @@ function ProjectsPage() {
             }
         };
         fetchProjects();
-    }, [])
+    }, []);
 
     return (
         <div className="flex flex-col min-h-screen items-center bg-(--prokress-beige-100)">
             <div className="w-90 md:w-2xl xl:w-7xl rounded-3xl mt-[5%] h-80 bg-(--prokress-beige-50) shadow-md">
                 <div className="bg-(--prokress-beige-0) rounded-t-3xl pt-2 pl-4 pb-2"><p className="text-(--prokress-black-700) text-2xl font-bold">Omat projektit</p></div>
-                <div className="flex flex-row gap-x-2 ml-4 mr-4 overflow-auto">
+                <div className="flex flex-row gap-x-2 ml-4 mr-4 overflow-auto scrollbar">
                     {myProjects.map((project, index) => {
                         return (
                             <div key={index}
@@ -72,7 +71,7 @@ function ProjectsPage() {
             </div>
             <div className="w-90 md:w-2xl xl:w-7xl rounded-3xl mt-[5%] h-80 bg-(--prokress-beige-50) shadow-md">
                 <div className="bg-(--prokress-beige-0) rounded-t-3xl pt-2 pl-4 pb-2"><p className="text-(--prokress-black-700) text-2xl font-bold">Jaetut projektit</p></div>
-                <div className="flex flex-row gap-x-2 ml-4 mr-4 overflow-auto">
+                <div className="flex flex-row gap-x-2 ml-4 mr-4 overflow-auto scrollbar">
                     {sharedProjects.map((project, index) => {
                         return (
                             <div key={index}
@@ -94,7 +93,6 @@ function ProjectsPage() {
                 </div>
             </div>
         </div>
-            {/* Project creation dialog */ }
     <ProjectDialogCreation
         isOpen={isProjectDialogOpen}
         toggleDialog={() => setProjectDialogOpen(false)}
